refactor: daily hero
